Another move towards library-based features
