+++ src/commands/global/cert.c
@@ -24,10 +24,10 @@
 #include "mbedtls/md.h"
 #include "mbedtls/pem.h"
 #include "pico/unique_id.h"
-#include "otp/bp_otp.h"
 #include "pubkey/pubkey.h"
 #include <boot/bootrom_constants.h>
 #include <pico/bootrom.h>
+#include "deps/saferotp/saferotp.h"
 
  
 #define BYTES_PER_OTP_PAGE 128
