+++ src/deps/saferotp/saferotp_ecc.c
@@ -1,4 +1,4 @@
-#include "bp_otp_ecc.h"
+#include "saferotp_ecc.h"
 
 
 static const uint32_t SUCCESS_MASK = 0x0000FFFFu;
